# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DFDR Swagger OpenAPI Documentation",
        "description": "Endpoint from DFDR Application",
        "contact": {
            "email": "admin@dfdr.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/get-notification": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "description": "View All",
                "operationId": "getNotification",
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/get-notification-approval/{userId}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "description": "View All",
                "operationId": "getNotificationApprovalUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            }
        },
        "/read-notification-user/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "description": "View All",
                "operationId": "readNotificationUser",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Notification User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            }
        },
        "/read-notification-upload/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "description": "View All",
                "operationId": "readNotificationUpload",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Notification Upload",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            }
        },
        "/read-notification-approval/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "description": "View All",
                "operationId": "readNotificationApproval",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Notification Approval",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            }
        },
...
        "/subject": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "description": "View All",
                "operationId": "findAllSubject",
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subject"
                ],
                "description": "Add new",
                "operationId": "postSubject",
                "requestBody": {
                    "description": "Add new Subject",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "subject_name"
                                ],
                                "properties": {
                                    "subject_name": {
                                        "description": "Subject Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subject/{id}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "description": "View By Id",
                "operationId": "showSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Subject",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subject"
                ],
                "description": "Update Data",
                "operationId": "putSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Subject",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subject_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Update Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subject"
                ],
                "description": "Delete Data",
                "operationId": "destroySubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Subject",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Delete Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subject/search/{name}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "description": "Search Data",
                "operationId": "searchSubject",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Subject",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Search Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer -token from login-)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}